readChunk in own module
